Move and rename billing periods service
Again, another service that will be generic amongst the bill run types. They might not use all the periods like supplementary but they do need the service to confirm what the current period is.

Whilst we're doing this we take the opportunity to rename it. `BillingPeriodsService` was the only one that didn't follow a `VerbNounService` naming convention. It's the little things!