Add shell 2PT process billing batch service
This will be the start point when we come to implement our 2PT bill run engine. Obviously we have nothing as yet so we just have it throw. But it means the generic stuff of creating the billing batch, billing batch event, the charging module bill run, and providing the appropriate response is all done ready (after a few more changes!)