Fix schema export service using streams
https://eaflood.atlassian.net/browse/WATER-4019

This PR's primary focus is to fix the schema export service. The service is currently not working due to the fact it reaches its memory capacity when it exports the returns table. This is due to the fact the full table is brought into memory. This fix introduced streams into the service to allow the table to be written immediately to a file and not be held in full in memory.