Create `ChargingModuleBillRunStatusService`
We found that we were getting intermittent errors which we suspect is due to us immediately requesting the invoice details after sending the reissue request; the CM is busy creating the new invoices and licences and therefore returns partial data. We therefore create a `ChargingModuleBillRunStatusService` which we can use to check that the bill run status is not `pending` before we continue.