Allow inserting of objects into transaction json fields
We were encountering errors when inserting transactions where their `purposes` fields were reported as being in the wrong format for a `jsonb` field. On investigation this is because Objection does not as standard allow you to insert an object into a jsonb field -- it needs to be stringified first. However Objection will handle this for us automatically if we define in the model which fields are json format using the `jsonAttributes` getter.