Fix reissuing ALCS invoice bug (#311)
https://eaflood.atlassian.net/browse/WATER-4036

Reissuing was confirmed working in our local environments but was continuing to fail in `dev`. When we checked the logs we realised that it was failing because it was fetching alcs invoices that were awaiting reissuing as well as sroc ones! This then causes an error to come back from the CM as it won't reissue an alcs invoice onto an sroc bill run.

We resolve this by updating `FetchInvoicesToBeReissuedService` to only fetch invoices linked to an sroc billing batch.