Identify Licences for 2 Part Tariff Annual Billing (SROC) (#324)
https://eaflood.atlassian.net/browse/WATER-4038

2 Part Tariff Billing is split over 2 Annual events, the 1st part is based on Authorised volumes linked to Spray or Trickle irrigation, where a 2 part tariff agreement is present and provides a 50% discount on those elements as part of the Annual Billing process.

The first step in the process that this PR will cover is to identify the licences that need to be included in the bill run.   This should be any licence within the chosen region which has a valid charge version within the previous year where the 2PT indicator at Charge Reference level is set to yes.  ie If the Bill run triggered is for FY 22/23 then the charge version must be valid in the FY 21/22.