Match CV data to returns data for 2PT billing
https://eaflood.atlassian.net/browse/WATER-4046

First iteration of mathcing returns to charge versions. This covers the first example which is a one-to-one match.

We also have updated the results to include charge versions, charge elements and charge purposes. We intend to refine this in future iterations but we are adding them to confirm the matching algorithm is working correctly.