Major version update of packages (#327)
The packages `got`, `hapi-pino` and `pino-pretty` are all one major version out of date. This PR will update them to the latest versions.

It is not possible to update the `got` package as v13 requires Node 16+ and we are running 14 on our servers.

Also whilst `dependabot` does minor & patch updates automatically it only updates the `package-lock.json`. So I've also tidied up the `package.json` to reflect the package updates already in the lock file.