Update migrations to follow team convention
As a team we have agreed that [our convention](https://github.com/DEFRA/water-abstraction-team/blob/main/coding_conventions.md#arrow-functions) when using arrow functions is to always wrap arguments in brackets even if there is only 1.

We spotted with a recent migration that we were not doing this. As we tend to copy and paste previous migrations when adding new ones we've decided to correct what we have to save repeating the infraction in future changes!