Update migrations to follow team convention (#330)
As a team, we have agreed that [our convention](https://github.com/DEFRA/water-abstraction-team/blob/main/coding_conventions.md#arrow-functions) when using arrow functions is to always wrap arguments in brackets even if there are only 1.

We spotted with a recent migration that we were not doing this. At the same time, we also clocked we were doing the following in some migrations.

```javascript
exports.up = async function (knex) {
  await knex
    .schema

// vs

exports.down = function (knex) {
  return knex
    .schema
```

The [Knex example](https://knexjs.org/guide/migrations.html#transactions-in-migrations) implies what we are doing in `down()` is correct. Looking at some of our more straightforward migrations, for example, when creating schemas we saw that's what we were already doing.

As we tend to copy and paste previous migrations when adding new ones we've decided to correct what we have to save repeating the infractions in future changes!