Fix broken process billing period unit test (#336)
We have found a unit test when run in isolation that breaks.

```
Process billing period service
  when the service is called
    and there are charge versions to process
      but none of them are billable
        because the charge version status is `superseded`
          and there are no previously billed transactions
            ✔ 326) returns false (bill run is empty) (5 ms)
```

It is just a copy & paste error. The `beforeEach()` block being used is missing a couple of lines of code that were overlooked when the block was copied.

Because `chargeVersions` is declared outside this test, as long as other tests run it is getting populated. But if the test is run in isolation it never gets set hence the error.