Create model for `purposes_uses` table (#335)
https://eaflood.atlassian.net/browse/WATER-4046

Whilst working on ticket 4046 we have realised that we need a model for the `purposes_uses` table to be able to match the Returns to a charge purpose. The Return has a value in the `metadata` that matches to the `legacy_id` in the `purposes_uses` table.

I also noticed while working on this model that we hadn't used the correct naming conventions for the file names of the `returns` schema migrations that we had recently built. Also in the migration `20230802134153_create-returns-returns` the constraints & defaults had been omitted so I also added these.