Add crm_v2 to DatabaseHelper clean() method (#337)
When we added the first `crm_v2` schema models to our repo we forgot to add the schema to our `DatabaseHelper` and its `clean()` method.

We use this to ensure the DB is empty of existing data before each test to ensure test isolation. This change adds the schema to the helper.

Note - we also had to tweak the query to wrap the schema name in quotes. It's because PostgreSQL doesn't actually like the name because it is case-sensitive 🤦😩. Without the quotes, everything falls over.

Co-authored-by: Stuart Adair <43574728+StuAA78@users.noreply.github.com>