Match returns to charge version by purpose type (#342)
https://eaflood.atlassian.net/browse/WATER-4046

We are now working on Example 2, a licence with a charge version with 4 elements. One has a single purpose. The other has an element with 3 purposes. We need to match 4 returns; 1 to the first element and the rest to the second.

This change links the charge purposes to returns via `water.purposes_uses`, specifically `legacy_id` to `metadata.purposes[0].tertiary.code`. It's this that allows us to match a return directly to a charge element.

We also realised that we were still assigning the returns at the wrong level. They needed to be assigned to the charge element they match not the charge version.

Finally, we spotted the `innerJoin()` in our `_fetchChargeVersions()` query was causing a charge version to appear more than once if it contained multiple charge elements.