Add friendly formatting to two-part tariff endpoint (#344)
https://eaflood.atlassian.net/browse/WATER-4046

We update our two-part tariff endpoint to output data in a friendlier format, optionally allowing us to view the raw data if specified.

The endpoint is still `/check/two-part/{region}` but it now takes an optional `/{format}` path parameter, eg `/check/two-part/1/raw`. We default to `friendly` output if nothing is specified, which is currently a simple `{ "hello": "world" }` response but will ultimately be in a format similar to the UI. `raw` format simply shows the end result of the db queries and matching without any further changes. Any other format value returns an error response.