Add timing to 2PT matching service (#346)
https://eaflood.atlassian.net/browse/WATER-4046

We believe we have the matching algorithm to a point that covers all examples we have been asked to meet so far. We're also confident it will meet any new ones thrown at us!

So, if and when we do make changes we want to make sure we retain its current performance (5 secs to match _all_ 2PT licences in Anglian; approximately 2,000).

To help us do that we are adding timing to the `TwoPartService` so we can keep an eye on it in the logs as we make further changes.