Add new data formatters to base presenter (#350)
https://eaflood.atlassian.net/browse/WATER-4070
https://eaflood.atlassian.net/browse/WATER-4073

Whilst working on [Create endpoint to generate fake data](https://github.com/DEFRA/water-abstraction-system/pull/347) we added a number of formatters. These take DB data and re-format it to match how the UI displays the values.

We will need these formatters in future changes we'll be making, the first of which is a re-designed bill run page.

So, to help prepare for that we are adding those formatters to our `BasePresenter` so they are available to any future presenters we add.