Use `server.dependency()` in plugin
Calling `server.dependency()` in a plugin allows to specify which we will wait to be registeted, and some code to be executed once it's registered. The rest of the plugin code is executed while waiting. In this case, we use `server.dependency()` to wait until `@hapi/cookie` is registered before we set up the authentication strategy, but immediately create the test route as this isn't dependent on anything else.

All of this ensures that the order of the plugins in our `registerPlugins()` function isn't important, ie. we have no issues if we register them in the "wrong" order. To demonstrate this, we deliberately register `@hapi/cookie` _after_ `AuthenticationPlugin`