Add new ErrorPagesService
Some boy scouting going on. Our `ErrorPagesPlugin` is currently doing a lot of work and normally we like to move the work into services.

It keeps the plugins simple and makes for easier testing. So, with this change requiring us to take another look at the plugin we are taking the opportunity to refactor the work into a service.

It is not a direct copy and paste because we learnt some things whilst testing the error handling.

- if it's an unhandled error, one that has been boomified or a 404 `response` will always we be a boom error. There is no point in trying to handle a situation where this is not the case
- we could stick with putting the `plainOutput:` option on the route's `options.settings.plugins` property. But as we are already using `options.settings.app` for custom config (see route filtering) it made sense to also use it for this

We also realised that we always want to log anything that is a 404, and send an airbrake notification of any other boom error so the service takes care of that.