Update plugin to use new service
We also tidy up the existing plugin tests. We remove some of the duplication with setting up the test route. We test non-erroring routes (2xx and 3xx) as the plugin needs to deal with those as well.

We also clarify the tests; we now know you wouldn't get a scenario where the response, if an error, is not a boom error.