Implement TwoPartService 'friendly' response
https://eaflood.atlassian.net/browse/WATER-4046

We have our `TwoPartService` which is our spike on coming up with a working two-part tariff matching algorithm. But the data its matching has issues

- there is _lots_ of it. The records have numerous fields and many are JSONB which means when presented they expand out even more
- for 'reasons' few of the tables and hence the entity names match what you see in the UI

Reducing the amount of data output to just the key elements will help when assessing a match is right. And as you'll generally have the UI open and the relevant record up it would _really_ help if the response matched what you saw on screen.

So, we have created a `FriendlyResponseService`. We fully intend to delete it once we have confirmation the matching algorithm is good. This is our excuse for why it looks horrendous and has no tests!

But it means the default response from our `/check` endpoint will make it a lot easier to assess the critical bit; the matching algorithm.