Multipart upload for aws bucket
https://eaflood.atlassian.net/browse/WATER-4086

During the testing of the data export service on the development environment, it was identified that the upload process was unsuccessful due to the excessive size of the files. The original approach attempted to upload the entire file in one go. To address this issue, this pull request introduces a modification to the upload process, enabling the file to be uploaded in multiple segments. This update has logic to assess the file size and determine the most suitable method for uploading. AWS restricts the application of multi-part uploads to files under 5MB (https://docs.aws.amazon.com/AmazonS3/latest/userguide/qfacts.html). Given this constraint, certain compresses files fall below the 5MB threshold and therefore cannot be uploaded through multi-part upload. This means its necessary to include both multi-part upload and single part.