Handle returns without alias in friendly response (#366)
https://eaflood.atlassian.net/browse/WATER-4046

We found in testing that some return records do not have an alias which we assumed all did. After checking what the UI does in these cases we found it just displays the description against the `nald.purposes[0].tertiary` object in the `metadata` as a fallback.

This updates the 'friendly' response to handle it.