Refactor controllers out of folders (#368)
https://eaflood.atlassian.net/browse/WATER-4092

This change is not directly related to WATER-4092. This is more of a technical debt exercise. We were about to embark on a spike of seeing if we could replicate what the legacy update billing account address code is doing. Repeating what we already have would mean creating a new folder to put the single controller file in.

This seems to be needlessly complicating the controllers' folder. It's not something we have done in other areas which highlights we are pre-empting future needs. Certainly, when a routes controller file becomes too unwieldy we will most likely break it down into separate controllers grouped in a folder.

But let's wait until we need to do that! 😁