Create `AuthenticationConfig` file
We originally read the `COOKIE_SECRET` environment variable directly so that we could quickly get a working plugin up and running. Now that we have it working, we create `authentication.config.js` in line with our usual pattern for accessing env vars.

Note that we only set the password in the env vars, not the cookie name, ttl etc. We do this because only the password is configurable in our other repos, so there's little point making other things configurable in this repo.