Add automatic payload cleaning
https://eaflood.atlassian.net/browse/WATER-4092

We really should have added this when working on supplementary billing. But now we are building a new endpoint to handle changing the address for a billing account we really need it.

We'll get into why we need to build a new endpoint when we start adding the code for it. But there are a series of changes we need to do in preparation.

First of these is to implement automatic payload cleaning. This is something we did in the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api). By cleaning payloads automatically when they come in, we avoid polluting the [Joi](https://joi.dev/) schemas and our controllers with lots of calls to `trim()` or checks to see if something is `null` or `undefined`.

By cleaning we mean removing empty, null or undefined properties then removing extraneous whitespace from the values left. It also handles sanitizing values to protect us from things like [cross site scripting](https://owasp.org/www-community/attacks/xss/).

With this in place, controllers can get on and use `request.payload` without having to do any of these other checks.