Add disinfect plugin to sanitize payloads
https://eaflood.atlassian.net/browse/WATER-4092

We really should have added this when working on supplementary billing. But now we are building a new endpoint to handle changing the address for a billing account we really need it.

We'll get into why we need to build a new endpoint when we start adding the code for it. But there are a series of changes we need to make in preparation.

The first of these is to implement automatic payload cleaning. This is something we did in the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api). By cleaning payloads automatically when they come in, we avoid polluting the [Joi](https://joi.dev/) schemas and our controllers with lots of calls to `trim()` or checks to see if something is `null` or `undefined`.

By cleaning we mean removing empty, null or undefined properties and then removing extraneous whitespace from the values left. It also includes sanitizing values to protect us from things like [cross-site scripting](https://owasp.org/www-community/attacks/xss/).

With this in place, controllers can get on and use `request.payload` without having to do any of these other checks.

In the **Charging Module** we rolled our own using [sanitizer](https://github.com/theSmaw/Caja-HTML-Sanitizer) to handle the dodgy stuff. @Jozzey pointed out [disinfect](https://github.com/genediazjr/disinfect) as an alternate. The fact it's been updated in the last 8 years already puts it ahead. But it can also do some of the work we were doing, dropping empty or null properties from the payload. The one thing it doesn't do is remove extraneous whitespace from values.

So, this change implements **Disinfect** along with a custom `payloadSanitizer` to get us the same result.