Add CRM_V2 company model
https://eaflood.atlassian.net/browse/WATER-4092

We are building a new endpoint to handle changing the address for a billing account.

We'll get into why we need to build a new endpoint when we start adding the code for it. But there are a series of changes we need to make in preparation.

The first of these was [Add automatic payload cleaning](https://github.com/DEFRA/water-abstraction-system/pull/375). The next set of changes is adding the models we'll need to interact with the `crm_v2` schema.