Add CRM_V2 contact model (#379)
https://eaflood.atlassian.net/browse/WATER-4092

We are building a new endpoint to handle changing the address for a billing account.

We'll get into why we need to build a new endpoint when we start adding the code for it. But there are a series of changes we need to make in preparation.

The first of these was [Add automatic payload cleaning](https://github.com/DEFRA/water-abstraction-system/pull/375). The next set of changes is adding the models we'll need to interact with the `crm_v2` schema.

The `contact` model includes some additional functionality to return the 'name' to use. Unfortunately, both the source NALD system and our existing legacy service have chosen to go all in on breaking apart a person's name. What's even better (!) is they do it in different ways!!

We'll eventually be trying to replicate existing functionality that currently has a bug in it. It depends on being able to generate a name. So, even though we do not agree with splitting apart names and capturing unnecessary details like titles and suffixes, we've had to implement our own version of what the legacy code was doing.

> For reference, the [GOV.UK Design system](https://design-system.service.gov.uk/patterns/names/) encourages you not to do what we currently do. You should also look at [Personal names around the world (W3C)](https://www.w3.org/International/questions/qa-personal-names) and [Falsehoods about names (Kalzumeus)](http://www.kalzumeus.com/2010/06/17/falsehoods-programmers-believe-about-names/) to understand why any pattern other than 'Name:' is bad!