Initial version and tests
The invalid tests just focus on those that involve checking a value is in one of our static lookups. Those checks in the schema depend on another area of our code so it felt worth while including tests for them.

We have a test to show that `address:` is a required property. Beyond that it felt like we'd just be testing Joi which is why every possible invalid data scenario is not covered.