Break schema down
TBH 50/50 on whether this improves things or not. We hope it makes the schemas for the individual properties easy to see and understand so we're going for it.

There is an argument though that seeing the whole schema in one place makes it easier to see what Joi is doing. ¯\_(ツ)_/¯