Create `group`, `role` and `user` models for authentication plugin (#387)
https://eaflood.atlassian.net/browse/WATER-4085

In anticipation of developing our [authentication plugin](#351) we create models for the `group`, `role` and `user` tables, along with helpers and unit tests.

The models are all straightforward ones with the exception of `UserModel`, which adds a static method `generateHashedPassword`. This takes a plaintext password and returns a salted and hashed version. The code is taken from `db/seeds/01-users.js`; the intent is that the seed file will at some point be refactored to insert users using `UserModel`, and therefore use the static method in place of the existing function.