Add "forward" relations to models
We want to replace our naive approach to retrieving the roles with something a little nicer. In order to do this, we need to add relations to the models so we can easily go from one to another. Note that for now we only add "forward" relations, eg. from user to user role to role. We don't yet add ones going the other way, eg. from role to user role to user. We also have not yet written unit tests for these.