Update notes for contact and company CRM_V2 models (#393)
https://eaflood.atlassian.net/browse/WATER-4092

We are working on migrating the change billing account address functionality from the legacy code to this project. Whilst working on persisting the company data we learnt some details about how the previous team stored data in the `crm_v2.companies` table depending on the source.

We wanted to add these notes for future reference, which made us revisit how we did something similar in the `ContactModel`.

So, this change is just about updating our notes in the models to be consistent.