Delete Create Billing Transaction Service
https://github.com/DEFRA/water-abstraction-team/issues/89

We have a simple service `app/services/supplementary-billing/create-billing-transaction.service.js` which exists to create transactions. We haven't actually used it anywhere, preferring instead to create them directly using `BillingTransactionModel`. This, coupled with the fact that we don't have corresponding services for creating billing invoices and billing invoice licences, means we should delete it.

It also allows us to clear up an ambiguity in the unit tests because the test file for this service is named differently. 😁